% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_command_found.R
\name{blast_command_found}
\alias{blast_command_found}
\title{Check BLAST Installation}
\usage{
blast_command_found(blast_command)
}
\arguments{
\item{blast_command}{String specifying the path to a BLAST program.}
}
\value{
Logical. Returns \code{TRUE} if the BLAST program could be found.
}
\description{
Checks whether a BLAST program can be found.
}
\examples{
blast_command_found(blast_command="blastn")
}
