% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_version.R
\name{blast_version}
\alias{blast_version}
\title{Get BLAST Version}
\usage{
blast_version(blast_command = "blastn")
}
\arguments{
\item{blast_command}{String specifying the path to a BLAST program. The default (\code{'blastn'}) should return the version of the blastn program for standard BLAST installations. The user can provide a path to a BLAST program for non-standard BLAST installations.}
}
\value{
Returns a string of the version of the BLAST program.
}
\description{
Gets the version of a BLAST program.
}
\examples{
blast_version()
}
