% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Draw Circle Polygon}
\usage{
circle(r, v = 1000, ...)
}
\arguments{
\item{r}{Numeric scalar of circle radius.}

\item{v}{Numeric scalar of vertex count (default = \code{1000}).}

\item{...}{Additional arguments passed to \code{\link[graphics:polygon]{polgyon}}.}
}
\value{
No return value.
}
\description{
Draws circle polygon.
}
\details{
Draws a circle polygon of a given radius. The circle is drawn about the origin (\emph{i.e.}, x = 0, y = 0). Intended for use with \code{\link[=template]{template}} to generate \code{\link[=detection]{detection}} and \code{\link{proportion}} plots.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
template(l=1)
circle(r=1)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=sector]{sector}} for plotting sector polygons.
}
