% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_wildcards.R
\name{contains_wildcards}
\alias{contains_wildcards}
\title{Check Whether DNA Sequences Contain Wildcard Characters}
\usage{
contains_wildcards(sequences)
}
\arguments{
\item{sequences}{A character vector of DNA sequences.}
}
\value{
A logical vector indicating whether each DNA sequence contains wildcard characters.
}
\description{
Checks whether DNA sequences contain wildcard characters.
}
\examples{
contains_wildcards(sequences=c("TKCTAGGTGW","CATAATTAGG","ATYGGCTATG"))
}
