% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Generate Circular Coordinates}
\usage{
coordinates(a, r)
}
\arguments{
\item{a}{Numeric vector of angles (degrees).}

\item{r}{Numeric scalar of the circle radius.}
}
\value{
A numeric matrix of xy coordinates.
}
\description{
Generates coordinates along a circular path.
}
\details{
Calculates xy coordinates along a circular path given a vector of angles and a specified radius. The center of the circle is at the origin (\emph{i.e.}, x = 0, y = 0). This function is helpful for calculating the vertices of circle and sector polygons.
}
\examples{
coordinates(a=c(90,180,270,360),r=1)
}
\seealso{
\code{\link[=circle]{circle}} for plotting circle polygons. \cr \cr
\code{\link[=sector]{sector}} for plotting sector polygons.
}
