% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2cov.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert Correlation to Covariance Matrix}
\usage{
cor2cov(sd, R)
}
\arguments{
\item{sd}{Numeric vector of standard deviations.}

\item{R}{Numeric correlation matrix.}
}
\value{
Returns a numeric covariance matrix.
}
\description{
Derives covariance matrix from correlation matrix and standard deviation vector.
}
\details{
Given correlation matrix \code{R} and standard deviation vector \code{sd}, performs the operation \code{\link[=diag]{diag}}\code{(sd)} \code{\link[=matmult]{\%*\%}} \code{R} \code{\link[=matmult]{\%*\%}} \code{\link[=diag]{diag}}\code{(sd)} to derive the corresponding covariance matrix. This is a counterpart to \code{\link[stats:cor]{stats::cov2cor}}, which scales a covariance matrix into the corresponding correlation matrix.
}
\examples{
# Define standard deviation vector.
sd<-c(9.655,1.157,1.128,2.925)

# Define correlation matrix.
R<-matrix(data=c(1.000,-0.80,0.64,-0.512,
                 -0.800,1.00,-0.80,0.640,
                 0.640,-0.80,1.00,-0.800,
                 -0.512,0.64,-0.80,1.000),
           ncol=4,byrow=TRUE)

# Derive covariance matrix.
cor2cov(sd=sd,R=R)
}
\seealso{
\code{\link[stats:cor]{stats::cov2cor}} for scaling a covariance matrix into the corresponding correlation matrix.
}
