% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcopula.R
\name{dcopula}
\alias{dcopula}
\title{Density of the Gaussian Copula}
\usage{
dcopula(u, R, log = FALSE)
}
\arguments{
\item{u}{Numeric vector or matrix of uniformly-distributed margins on the interval [\code{0}, \code{1}]. If matrix, then a vector of probability densities is returned with an element for each record of the matrix. Matrix records represent observations, and matrix fields represent dimensions.}

\item{R}{Numeric correlation matrix. If \code{u} is a matrix, then \code{R} is recycled for each record of matrix \code{u}.}

\item{log}{Logical scalar. If \code{TRUE}, then probabilities are given as \code{log(density)}.}
}
\value{
Numeric vector of probability densities.
}
\description{
Density function for the Gaussian copula.
}
\details{
Computes the probability density of the Gaussian copula. Given uniformly-distributed margins \code{u} on the interval [\code{0}, \code{1}], applies the inverse cumulative distribution function of the standard normal (\emph{i.e.}, \code{\link[stats:Normal]{stats::qnorm}}) to map uniform margins to normal scores. Then, uses equation 1 of Song (2000) with the normal scores to calculate the probability density of the Gaussian copula.
}
\examples{
# Define uniform margins.
u<-c(0.324,0.383,0.917,0.015)

# Define correlation matrix.
R<-matrix(data=c(1.000,-0.80,0.64,-0.512,
                 -0.800,1.00,-0.80,0.640,
                 0.640,-0.80,1.00,-0.800,
                 -0.512,0.64,-0.80,1.000),
           ncol=4,byrow=TRUE)

# Compute log probability density.
dcopula(u=u,R=R,log=TRUE)
}
\references{
Song P. 2000. Multivariate dispersion models generated from Gaussian copula. \emph{Scandinavian Journal of Statistics}, 27(2): 305-320. DOI: 10.1111/1467-9469.00191
}
\seealso{
\code{\link[=dmvlogis]{dmvlogis}} for density of the multivariate logistic distribution.
}
