% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_quality_scores.R
\name{decode_quality_scores}
\alias{decode_quality_scores}
\title{Decode DNA Sequence Quality Scores}
\usage{
decode_quality_scores(symbols)
}
\arguments{
\item{symbols}{A string containing quality scores encoded as symbols in Sanger format.}
}
\value{
A numeric vector of Phred quality scores.
}
\description{
Decodes Phred quality scores in Sanger format from symbols to numeric values.
}
\examples{
decode_quality_scores(symbols="989!.C;F@\"")
}
