% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detection}
\alias{detection}
\title{Grouped Detection Plot}
\usage{
detection(
  x,
  r = 1,
  b = 0.025,
  v = 1000,
  w = 1,
  f = 0.5,
  c = "lightskyblue",
  m = 3,
  ...
)
}
\arguments{
\item{x}{A list of vectors named \code{"g"}, \code{"s"}, \code{"r"}, and \code{"d"}. The elements of vector \code{"g"} (character, numeric, or factor) specify the group. The elements of vector \code{"s"} (character or numeric) specify the sample. The elements of vector \code{"r"} (numeric) specify the number of replicates within sample \code{"s"}. The elements of vector \code{"d"} (numeric) specify the number of replicates within sample \code{"s"} with detections.}

\item{r}{Numeric scalar. Radius of plot circle (default = \code{1}).}

\item{b}{Numeric scalar. Plot radius buffer (proportion; default = \code{0.025}).}

\item{v}{Numeric scalar. Vertex count of plot circle (default = \code{1000}).}

\item{w}{Numeric scalar. Line width of outer circle (default = \code{1}).}

\item{f}{Numeric scalar. Line width of sectors as a proportion of \code{w} (default = \code{0.5}).}

\item{c}{Character string. Fill color of sub-sector detections (default = \code{"lightskyblue"}).}

\item{m}{Numeric scalar. Maximum number of plot columns (default = \code{3}).}

\item{...}{Additional arguments passed to \code{\link[graphics:title]{title}}.}
}
\value{
No return value.
}
\description{
Generates detection plots for multiple groups.
}
\details{
Produces a pie-chart-like detection plot with grouping structure. Each circle represents a group. Each sector represents a sample, and each sub-sector represents a replicate. Filled replicates represent detections. Groups are sorted alphabetically (or inherit factor level ordering) and arranged from left to right and top to bottom. Samples are sorted alphabetically and arranged in a clockwise orientation (from angle zero). Samples are sorted independently for each group. This plot design is specialized for visualizing binary detection data.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
set.seed(1234)
n.groups<-6
n.samples<-6
n.replicates<-3
data<-list(g=rep(x=LETTERS[1:n.groups],each=n.samples),
           s=rep(x=letters[1:n.samples],times=n.groups),
           r=rep(x=n.replicates,times=n.groups*n.samples),
           d=sample(x=0:n.replicates,size=n.groups*n.samples,
                    replace=TRUE))
detection(x=data)
\dontshow{\}) # examplesIf}
}
\references{
A manuscript describing this plot design is in preparation.
}
\seealso{
\code{\link[=singular.detection]{singular.detection}} for singular detection plots. \cr \cr
\code{\link[=proportion]{proportion}} for grouped proportion plots.
}
