% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Bray-Curtis Dissimilarity (Proportion Form)}
\usage{
dissimilarity(p1, p2)
}
\arguments{
\item{p1}{Numeric vector or matrix of proportional abundances for first community. See details.}

\item{p2}{Numeric vector or matrix of proportional abundances for second community. See details.}
}
\value{
Numeric scalar or vector of Bray-Curtis dissimilarity values.
}
\description{
Compute Bray-Curtis dissimilarity from proportional abundances.
}
\details{
Calculates Bray-Curtis dissimilarity from proportional abundances using the formula \code{\link[=sum]{sum}}\code{(}\code{\link[=abs]{abs}}\verb{(p1-p2))/2}. This is equivalent to the Bray-Curtis dissimilarity formula in the \code{vegdist} function of the \code{vegan} package when both communities have the same total counts (\emph{e.g.}, rarefied counts). This function is primarily intended to provide a method to compute Bray-Curtis dissimilarity from the posterior predictions of Dirichlet-multinomial regression models, which generate predictions of proportional abundances.

The dimensions of \code{p1} and \code{p2} must match. If \code{p1} and \code{p2} are matrices, then each record represents paired replicates for the communities (\emph{e.g.}, predictions from the same posterior draw). The elements (if \code{p1} and \code{p2} are vectors) or fields (if \code{p1} and \code{p2} are matrices) represent dimensions (\emph{e.g.}, taxa or species). If \code{p1} and \code{p2} are vectors, then a numeric scalar is returned. If \code{p1} and \code{p2} are matrices, then a numeric vector is returned whose elements correspond to the paired records of \code{p1} and \code{p2}.
}
\examples{
# Compute Bray-Curtis dissimilarity.
dissimilarity(p1=c(0.15,0.25,0.4,0.2),
              p2=c(0.25,0.35,0.1,0.3))
}
\references{
Bray JR, and Curtis JT. 1957. An ordination of the upland forest communities of southern Wisconsin. \emph{Ecological Monographs}, 27(4): 325-349. DOI: 10.2307/1942268 \cr \cr
Legendre P, and Legendre L. 2012. \emph{Numerical Ecology: Third Edition}. Elsevier. \cr \cr
Odum EP. 1950. Bird populations of the Highlands (North Carolina) Plateau in relation to plant succession and avian invasion. \emph{Ecology}, 31(4): 587-605. DOI: 10.2307/1931577
}
\seealso{
\code{\link[=dmreg]{dmreg}} for fitting Dirichlet-multinomial regression models. \cr \cr
\code{\link[=dmpredict]{dmpredict}} for generating predictions from Dirichlet-multinomial regression models. \cr \cr
\code{\link[=diversity]{diversity}} for computing Hill diversity from proportional abundances. \cr \cr
\code{\link[=richness]{richness}} for computing species richness from occupancy probabilities.
}
