% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Hill Diversity}
\usage{
diversity(p, alpha = 2)
}
\arguments{
\item{p}{Numeric vector or matrix of proportional abundances. If vector, then Hill diversity is computed for the vector of proportions (and a scalar is returned). If matrix, then Hill diversity is computed independently for each record (and a vector is returned).}

\item{alpha}{Numeric scalar or vector. Continuous positive alpha parameter of the Hill diversity formula. If scalar, then \code{alpha} is recycled for each record of matrix \code{p}. If vector, then each element of \code{alpha} is applied to the corresponding record of matrix \code{p}. With the default of \code{alpha = 2}, Hill diversity is equal to the inverse Simpson index.}
}
\value{
Numeric scalar or vector of Hill diversity values.
}
\description{
Compute Hill diversity from proportional abundances.
}
\details{
Calculates Hill diversity from proportional abundances as defined in Hill (1973), which provides a unifying theory for ecological diversity indices. When \code{alpha = 0}, Hill diversity is equal to species richness. When \code{alpha = 1}, Hill diversity is equal to the exponentiated Shannon's entropy. When \code{alpha = 2} (the default), Hill diversity is equal to the inverse of Simpson's index. For any value of \code{alpha}, the Hill diversity of a community with uniform proportional abundances is equal to species richness. Hill diversity represents the effective number of species.
}
\examples{
# Compute Hill diversity.
diversity(p=c(0.15,0.25,0.4,0.2))
}
\references{
Hill MO. 1973. Diversity and evenness: A unifying notation and its consequences. \emph{Ecology}, 54(2): 427-432. DOI: 10.2307/1934352
}
\seealso{
\code{\link[=dissimilarity]{dissimilarity}} for computing Bray-Curtis dissimilarity from proportional abundances. \cr \cr
\code{\link[=richness]{richness}} for computing species richness from occupancy probabilities.
}
