% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmpredict.R
\name{dmpredict}
\alias{dmpredict}
\title{Predictions for Dirichlet-Multinomial Regression Models}
\usage{
dmpredict(X, H, fit, names)
}
\arguments{
\item{X}{Numeric predictor matrix. Predictions are made for each record. Each field represents a predictor variable, and the predictor variables must match (in order) those used to fit the \code{\link[=dmreg]{dmreg}} model. Matrix cells contain predictor values. Element names in the returned list are taken from the row names of \code{X}.}

\item{H}{Numeric vector or matrix (optional). If provided, then hierarchical effects are included in the predictions. Vector or matrix elements contain integer identifiers for values of hierarchical variables. If vector, then a single hierarchical variable is included, with each element corresponding to a record in \code{X}. If matrix, then each record in \code{H} corresponds to a record in \code{X}. Each field in \code{H} represents a hierarchical variable, and the hierarchical variables must match (in order) those used to fit the \code{\link[=dmreg]{dmreg}} model. If \code{H} is omitted, then hierarchical effects are omitted from the predictions (but may still have been used to fit the model).}

\item{fit}{A \code{stanfit} object returned from the \code{\link[=dmreg]{dmreg}} function. The fitted Dirichlet-multinomial regression model.}

\item{names}{Vector (optional). If provided, then field names in the matrices of the returned list will receive these values. If omitted, then the matrices in the returned list will lack field names.}
}
\value{
A list whose elements contain numeric matrices of posterior predictions. Within the list, one element is returned for each record of \code{X}. Element names are taken from the row names of \code{X}.
}
\description{
Generate predictions for Dirichlet-multinomial regression models.
}
\details{
Generates posterior predictions for Dirichlet-multinomial regression models fit with the \code{\link[=dmreg]{dmreg}} function. Predictions can either include or omit hierarchical effects, depending on whether argument \code{H} is provided. Returns a list where each element contains a matrix of posterior predictions for the respective record of \code{X}. Field names for the element matrices can optionally be provided with the \code{names} argument.
}
\examples{
# Define example data file path.
path<-system.file("extdata",
                  "example_regression_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Predict with fitted Dirichlet-multinomial regression.
out<-dmpredict(X=data$X,fit=data$fit,names=colnames(data$Y))
}
\seealso{
\code{\link[=dmreg]{dmreg}} for fitting Dirichlet-multinomial regression models. \cr \cr
\code{\link[=dmWAIC]{dmWAIC}} for computing widely applicable information criteria for Dirichlet-multinomial regression models.
}
