% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvlogis.R
\name{dmvlogis}
\alias{dmvlogis}
\title{Density of the Multivariate Logistic Distribution}
\usage{
dmvlogis(x, location, scale, R, log = FALSE)
}
\arguments{
\item{x}{Numeric vector or matrix. Values of logistically-distributed marginals. If matrix, then a vector of probability densities is returned with an element for each record of the matrix. Matrix records represent observations, and matrix fields represent dimensions.}

\item{location}{Numeric vector. Location parameters of the logistic distribution. If \code{x} is a matrix, then \code{location} is recycled for each record of matrix \code{x}.}

\item{scale}{Numeric vector. Scale parameters of the logistic distribution. If \code{x} is a matrix, then \code{scale} is recycled for each record of matrix \code{x}.}

\item{R}{Numeric correlation matrix. If \code{x} is a matrix, then \code{R} is recycled for each record of matrix \code{x}.}

\item{log}{Logical scalar. If \code{TRUE}, then probabilities are given as \code{log(density)}.}
}
\value{
Numeric vector of probability densities.
}
\description{
Density function for the multivariate logistic distribution.
}
\details{
Computes the probability density of the multivariate logistic distribution. The multivariate logistic distribution is constructed using a Gaussian copula with logistic marginals. The probability density is the product of the densities of the logistic marginals, which is further multiplied by the density of a Gaussian copula of the transformed standard uniform margins (\emph{i.e.}, probability integral transformation of the logistic marginals with \code{\link[stats:Logistic]{stats::plogis}}).
}
\examples{
# Define logistic margins.
x<-c(0.055,-1.625,0.329,-5.765)

# Define location parameters.
location<-c(0.477,-0.998,-0.776,0.064)

# Define scale parameters.
scale<-c(0.574,1.314,0.460,1.393)

# Define correlation matrix.
R<-matrix(data=c(1.000,-0.80,0.64,-0.512,
                 -0.800,1.00,-0.80,0.640,
                 0.640,-0.80,1.00,-0.800,
                 -0.512,0.64,-0.80,1.000),
           ncol=4,byrow=TRUE)

# Compute log probability density.
dmvlogis(x=x,location=location,
         scale=scale,R=R,
         log=TRUE)
}
\references{
Decani JS, and Stine RA. 1986. A note on deriving the information matrix for a logistic distribution. \emph{The American Statistician}, 40(3): 220-222. DOI: 10.2307/2684541  \cr \cr
Song P. 2000. Multivariate dispersion models generated from Gaussian copula. \emph{Scandinavian Journal of Statistics}, 27(2): 305-320. DOI: 10.1111/1467-9469.00191
}
\seealso{
\code{\link[stats:Logistic]{stats::dlogis}} for density of the logistic distribution. \cr \cr
\code{\link[=dcopula]{dcopula}} for density of the Gaussian copula.
}
