% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmWAIC.R
\name{dmWAIC}
\alias{dmWAIC}
\title{WAIC for Dirichlet-Multinomial Regression Models}
\usage{
dmWAIC(
  Y,
  X,
  H,
  ones = TRUE,
  method = 2,
  priors = c(B.mu = 0, B.sd = 1, theta.mu = 0, theta.sd = 1, sigma2.alpha = 0.01,
    sigma2.beta = 0.01),
  control = list(adapt_delta = 0.95, max_treedepth = 20),
  ...
)
}
\arguments{
\item{Y}{Numeric response matrix. Each record represents an observation, and each field represents a response dimension. Matrix cells contain integer counts.}

\item{X}{Numeric predictor matrix. Each record represents an observation, and each field represents a predictor variable. Matrix cells contain predictor values.}

\item{H}{Numeric vector or matrix (optional). If provided, then hierarchical effects are included in the model. Vector or matrix elements contain integer identifiers for values of hierarchical variables. If vector, then a single hierarchical variable is included, with each element representing an observation. If matrix, then each record represents an observation, and each field represents a hierarchical variable. Up to four hierarchical variables are supported (each with an arbitrary number of hierarchical levels).}

\item{ones}{Logical scalar. If \code{TRUE} (the default), then one is added to each cell of the response matrix. This avoids numerical errors which occur when distributional parameters in the model approach zero. For more information, see Harrison \emph{et al.} (2020). If the response matrix contains no zeros, then \code{ones} may be set to \code{FALSE}.}

\item{method}{Numeric scalar. Options are \code{1} or \code{2}, representing the alternative WAIC bias correction formulas (\emph{p}WAIC1 and \emph{p}WAIC2, respectively) described in Gelman \emph{et al.} (2014). As recommended by Gelman \emph{et al.} (2014), the default \code{method} (\code{2}) uses the \emph{p}WAIC2 bias correction formula.}

\item{priors}{Named numeric vector. Elements represent the prior values of their respective named parameters. When predictors are centered and scaled, the defaults generally represent weakly informative priors. Regression coefficients (\code{B}) and the precision parameter (\code{theta}) receive normal priors (with standard normal as the default). If hierarchical variables (argument \code{H}) are provided, then the common variances receive inverse-gamma priors (with default \code{alpha} and \code{beta} parameters of \code{0.01}).}

\item{control}{Named list of parameters which control the behavior of the Stan sampler. Passed to the \code{control} argument of the \code{rstan::sampling} function.}

\item{...}{Additional arguments passed to the \code{rstan::sampling} function.}
}
\value{
Returns numeric scalar of the widely applicable information criterion.
}
\description{
Computes the widely applicable information criterion (WAIC) for Dirichlet-multinomial regression models. Serves as a wrapper for \code{\link[=dmreg]{dmreg}}, \code{\link[=dmpredict]{dmpredict}}, \code{\link[=ddirmult]{ddirmult}}, and \code{\link[=waic]{waic}} for convenient WAIC calculations. Installation of the \code{rstan} package is required to use this function.
}
\details{
For convenience, wraps the steps involved in WAIC calculations for Bayesian Dirichlet-multinomial regression models. Begins by fitting a Bayesian Dirichlet-multinomial regression model with the \code{\link[=dmreg]{dmreg}} function, then generates resubstitution posterior predictions using the \code{\link[=dmpredict]{dmpredict}} function. The pointwise log-likelihood is calculated with the \code{\link[=ddirmult]{ddirmult}} function given the response matrix, posterior predictions, and precision parameter. WAIC is calculated from the pointwise log-likelihood using the \code{\link[=waic]{waic}} function.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Define example data file path.
path<-system.file("extdata",
                  "example_regression_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Compute WAIC for Dirichlet-multinomial regression.
out<-dmWAIC(Y=data$Y,X=data$X,H=data$H)
\dontshow{\}) # examplesIf}
}
\references{
Carpenter B, Gelman A, Hoffman MD, Lee D, Goodrich B, Betancourt M, Brubaker M, Guo J, Li P, and Riddell A. 2017. Stan: A probabilistic programming language. \emph{Journal of Statistical Software}, 76: 1-32. DOI: 10.18637/jss.v076.i01 \cr \cr
Gelman A, Hwang J, and Vehtari A. 2014. Understanding predictive information criteria for Bayesian models. \emph{Statistics and Computing}, 24(6): 997-1016. DOI: 10.1007/s11222-013-9416-2 \cr \cr
Goodwin KB, Hutchinson JD, and Gompert Z. 2022. Spatiotemporal and ontogenetic variation, microbial selection, and predicted \emph{Bd}-inhibitory function in the skin-associated microbiome of a Rocky Mountain amphibian. \emph{Frontiers in Microbiology}, 13: 1020329. DOI: 10.3389/fmicb.2022.1020329 \cr \cr
Harrison JG, Calder WJ, Shastry V, and Buerkle CA. Dirichlet-multinomial modelling outperforms alternatives for analysis of microbiome and other ecological count data. \emph{Molecular Ecology Resources}, 20(2): 481-497. DOI: 10.1111/1755-0998.13128 \cr \cr
Watanabe S. 2010. Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory. \emph{Journal of Machine Learning Research}, 11(116): 3571-3594.
}
\seealso{
\code{\link[=dmreg]{dmreg}} for fitting Dirichlet-multinomial regression models. \cr \cr
\code{\link[=dmpredict]{dmpredict}} for generating predictions from Dirichlet-multinomial regression models. \cr \cr
\code{\link[=ddirmult]{ddirmult}} for probability mass function of the Dirichlet-multinomial distribution. \cr \cr
\code{\link[=waic]{waic}} for generic function to compute widely applicable information criterion.
}
