% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_taxonomies.R
\name{expand_taxonomies}
\alias{expand_taxonomies}
\title{Expand Taxonomies}
\usage{
expand_taxonomies(
  taxonomies,
  levels = c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  full_names = TRUE,
  delimiter = ";",
  ignore
)
}
\arguments{
\item{taxonomies}{A character vector of taxonomic strings.}

\item{levels}{A character vector of taxonomic level names. The length of \code{levels} determines the number of taxonomic levels to extract from the taxonomies, and the order of elements in \code{levels} is assumed to match the order of taxonomic levels in the taxonomies. \code{levels} is also used as the field names of the returned data frame (see return value section). The default vector includes: \code{"Domain"}, \code{"Phylum"}, \code{"Class"}, \code{"Order"}, \code{"Family"}, \code{"Genus"}, and \code{"Species"}.}

\item{full_names}{Logical. If \code{TRUE} (the default), then full taxonomies are returned down to the extracted taxonomic level. If \code{FALSE}, then only the extracted taxonomic level is returned.}

\item{delimiter}{A character string of the delimiter between taxonomic levels in the input taxonomies. The default is \code{";"}.}

\item{ignore}{An optional character vector of taxonomic strings for which taxonomic expansion will be skipped. In the returned data frame (see return value section), the record for each skipped taxonomic string will be filled with \code{NA}s.}
}
\value{
Returns a data frame of extracted taxonomic levels. One record for each element of \code{taxonomies}, and one field for each element of \code{levels}. Field names are inherited from \code{levels}. If a taxonomic level is not present in a taxonomic string, then the respective cell in the returned data frame will contain \code{NA}.
}
\description{
Extracts each taxonomic level from a vector of taxonomic strings.
}
\examples{
print("Insert example here.")
}
\seealso{
\code{\link[=get_taxonomic_level]{get_taxonomic_level}} for extracting a taxonomic level from taxonomic strings. \cr \cr
\code{\link[=get_consensus_taxonomy]{get_consensus_taxonomy}} for generating a consensus taxonomy from taxonomic strings.
}
