% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{filter_sequences}
\alias{filter_sequences}
\title{Filter DNA Sequences by PCR Replicates}
\usage{
filter_sequences(
  input_files,
  samples,
  PCR_replicates,
  output_file,
  minimum_reads.PCR_replicate = 1,
  minimum_reads.sequence = 1,
  minimum_proportion.sequence = 0.005,
  binomial_test.enabled = TRUE,
  binomial_test.p.adjust.method = "none",
  binomial_test.alpha_level = 0.05,
  minimum_PCR_replicates = 2,
  delimiter.read_counts = ": ",
  delimiter.PCR_replicates = ", "
)
}
\arguments{
\item{input_files}{A character vector of file paths to input FASTA files. DNA sequences in the input FASTA files are assumed to be summarized by frequency of occurrence, with each FASTA header line beginning with "Frequency: " and followed by the sequence's read count. Output FASTA files from \code{\link[=truncate_and_merge_pairs]{truncate_and_merge_pairs}} have this format and can be used directly with this function. Each input FASTA file is assumed to contain the DNA sequence reads for a single PCR replicate for a single sample.}

\item{samples}{A character vector of sample identifiers, with one element for each element of \code{input_files}.}

\item{PCR_replicates}{A character vector of PCR replicate identifiers, with one element for each element of \code{input_files}.}

\item{output_file}{String specifying path to output file of filtered sequences in CSV format.}

\item{minimum_reads.PCR_replicate}{Numeric. PCR replicates which contain fewer reads than this value are discarded and do not contribute detections to any sequence. The default is \code{1} (\emph{i.e.}, no PCR replicates discarded).}

\item{minimum_reads.sequence}{Numeric. For a sequence to be considered detected within a PCR replicate, the sequence's read count within the PCR replicate must match or exceed this value. The default is \code{1} (\emph{i.e.}, no filtering by minimum read count within PCR replicates).}

\item{minimum_proportion.sequence}{Numeric. For a sequence to be considered detected within a PCR replicate, the proportion of reads in the PCR replicate comprised by the sequence must exceed this value. If \code{binomial_test.enabled = TRUE}, then this argument is used as the null hypothesis for a one-sided binomial test, and a significance test is used to determine whether the minimum proportion requirement for detection is satisfied instead. See the \code{binomial_test.enabled} argument below. The default is \code{0.005} (\emph{i.e.}, 0.5\%). To disable sequence filtering by minimum proportion within PCR replicates, set to \code{0}.}

\item{binomial_test.enabled}{Logical. If \code{TRUE} (the default), then for a sequence to be considered detected within a PCR replicate, the proportion of reads in the PCR replicate comprised by the sequence must significantly exceed the value of the \code{minimum_proportion.sequence} argument at the provided alpha level (\code{binomial_test.alpha_level} argument) based on a one-sided binomial test (\emph{i.e.}, \code{\link[=binomial_test]{binomial_test}} with \code{alternative = "greater"}). Optionally, p-values within a PCR replicate can be adjusted for multiple hypothesis testing by setting the \code{binomial_test.p.adjust.method} argument below. To disable significance testing, set to \code{FALSE} (minimum proportion filtering will still occur if \code{minimum_proportion.sequence > 0}, see above).}

\item{binomial_test.p.adjust.method}{String specifying the p-value adjustment method for multiple hypothesis testing. p-value adjustments are performed within each PCR replicate for each sample. Passed to the \code{method} argument of \code{\link[stats:p.adjust]{p.adjust}} in the \code{\link[stats:stats-package]{stats}} package. Available methods are contained within the \code{stats::p.adjust.methods} vector. If \code{"none"} (the default), then p-value adjustments are not performed. Ignored if \code{binomial_test.enabled = FALSE}.}

\item{binomial_test.alpha_level}{Numeric. The alpha level used in deciding whether the proportion of reads in a PCR replicate comprised by a sequence significantly exceeds a minimum threshold required for detection. See the \code{binomial_test.enabled} argument. The default is \code{0.05}. Ignored if \code{binomial_test.enabled = FALSE}.}

\item{minimum_PCR_replicates}{Numeric. The minimum number of PCR replicates in which a sequence must be detected in order to be considered present in a sample (\emph{i.e.}, not erroneous). The default is \code{2}.}

\item{delimiter.read_counts}{String specifying the delimiter between PCR replicate identifiers and sequence read counts in the Read_count_by_PCR_replicate field of the output CSV file (see details section). The default is \code{": "}.}

\item{delimiter.PCR_replicates}{String specifying the delimiter between PCR replicates in the Read_count_by_PCR_replicate field of the output CSV file (see details section). The default is \code{", "}.}
}
\value{
Invisibly returns a data frame containing unfiltered DNA sequences with their read counts, proportions, and p-values in each PCR replicate. While the primary output of this function is the written CSV file of filtered sequences described in the details section, the invisibly returned data frame of unfiltered sequences can be helpful when calibrating or troubleshooting filtering parameters. To aid in troubleshooting filtering parameters, the data frame is invisibly returned even if the error "Filtering removed all sequences" is received. Field definitions for the invisibly returned data frame of unfiltered sequences are:
\itemize{
\item Sample: The sample name.
\item PCR_replicate: The PCR replicate identifier.
\item Sequence: The DNA sequence.
\item Read_count.sequence: The sequence's read count within the PCR replicate.
\item Read_count.PCR_replicate: The number of reads in the PCR replicate.
\item Proportion_of_PCR_replicate.observed: The proportion of reads in the PCR replicate comprised by the sequence.
\item Proportion_of_PCR_replicate.null (Field only present if \code{binomial_test.enabled = TRUE}): The null hypothesis for a one-sided binomial test (inherited from the \code{minimum_proportion.sequence} argument). See the p.value field below.
\item p.value (Field only present if \code{binomial_test.enabled = TRUE}): The p-value from a one-sided binomial test of whether the proportion of reads in the PCR replicate comprised by the sequence exceeds the null hypothesis (\emph{i.e.}, \code{\link[=binomial_test]{binomial_test}} with \code{alternative = "greater"}).
\item p.value.adjusted (Field only present if \code{binomial_test.enabled = TRUE}): The p-value from the one-sided binomial test adjusted for multiple comparisons within each PCR replicate for each sample. See the p.value_adjustment_method field below.
\item p.value_adjustment_method (Field only present if \code{binomial_test.enabled = TRUE}): The p-value adjustment method (inherited from the \code{binomial_test.p.adjust.method} argument).
}
}
\description{
Filters DNA sequences by minimum read count within a PCR replicate, minimum proportion within a PCR replicate, and number of detections across PCR replicates.
}
\details{
For each set of input polymerase chain reaction (PCR) replicate FASTA files associated with a sample, writes out DNA sequences which are detected across a minimum number of PCR replicates (\code{minimum_PCR_replicates} argument). Detection within a PCR replicate is defined as a sequence having at least a minimum read count \emph{and} exceeding a minimum proportion of reads (\code{minimum_reads.sequence} and \code{minimum_proportion.sequence} arguments, respectively). When \code{binomial_test.enabled = TRUE}, a sequence must significantly exceed the minimum proportion within a PCR replicate at the provided alpha level (\code{binomial_test.alpha_level} argument) based on a one-sided binomial test (\emph{i.e.}, \code{\link[=binomial_test]{binomial_test}} with \code{alternative = "greater"}). Within a PCR replicate, p-values can be adjusted for multiple hypothesis testing by setting the \code{binomial_test.p.adjust.method} argument (see \code{stats::p.adjust.methods} and \code{\link[stats:p.adjust]{p.adjust}} in the \code{\link[stats:stats-package]{stats}} package). PCR replicates which contain fewer than a minimum number of reads are discarded (\code{minimum_reads.PCR_replicate} argument) and do not contribute detections to any sequence. \cr \cr
DNA sequences in the input FASTA files are assumed to be summarized by frequency of occurrence, with each FASTA header line beginning with "Frequency: " and followed by the sequence's read count. Output FASTA files from \code{\link[=truncate_and_merge_pairs]{truncate_and_merge_pairs}} have this format and can be used directly with this function. Each input FASTA file is assumed to contain the DNA sequence reads for a single PCR replicate for a single sample. \cr \cr
For pipeline calibration purposes, a data frame containing unfiltered DNA sequences with their read counts, proportions, and p-values in each PCR replicate is invisibly returned (see return value section). While the primary output of this function is the written CSV file of filtered sequences (described below), the invisibly returned data frame of unfiltered sequences can be helpful when calibrating or troubleshooting filtering parameters. To aid in troubleshooting filtering parameters, the data frame is invisibly returned even if the error "Filtering removed all sequences" is received. \cr \cr
For the primary output, this function writes a CSV file of filtered DNA sequences with the following field definitions:
\itemize{
\item Sample: The sample name.
\item Sequence: The DNA sequence.
\item Detections_across_PCR_replicates: The number of PCR replicates the sequence was detected in.
\item Read_count_by_PCR_replicate: The sequence's read count in each PCR replicate the sequence was detected in.
\item Sequence_read_count: The sequence's total read count across the PCR replicates the sequence was detected in. Calculated as the sum of the read counts in the Read_count_by_PCR_replicate field.
\item Sample_read_count: The sample's total read count across all sequences detected in the PCR replicates. Calculated as the sum of the read counts in Sequence_read_count field associated with the sample.
\item Proportion_of_sample: The proportion of sample reads comprised by the sequence. Calculated by dividing the Sequence_read_count field by the Sample_read_count field. Equivalent to the weighted average of the sequence's proportion in each PCR replicate, with weights given by the proportion of the sample's total reads contained in each PCR replicate.
}
}
\examples{
print("Insert example here.")
}
\references{
A manuscript describing these methods is in preparation.
}
\seealso{
\code{\link[=binomial_test]{binomial_test}} for performing vectorized one-sided binomial tests. \cr \cr
\code{\link[=truncate_and_merge_pairs]{truncate_and_merge_pairs}} for truncating and merging read pairs prior to sequence filtering. \cr \cr
\code{\link[=local_taxa_tool]{local_taxa_tool}} for performing geographically-conscious taxonomic assignment of filtered sequences.
}
