% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomies.IUCN.R
\name{get_taxonomies.IUCN}
\alias{get_taxonomies.IUCN}
\title{Get Taxonomies from IUCN Red List Files}
\usage{
get_taxonomies.IUCN(
  path_to_IUCN_taxonomies,
  path_to_IUCN_common_names,
  path_to_output_local_taxa_list,
  domain_name = "Eukaryota",
  path_to_taxonomy_edits = NA
)
}
\arguments{
\item{path_to_IUCN_taxonomies}{String specifying path to input IUCN Red List taxonomy.csv file.}

\item{path_to_IUCN_common_names}{String specifying path to input IUCN Red List common_names.csv file.}

\item{path_to_output_local_taxa_list}{String specifying path to output species list (in CSV format with '.csv' extension) with formatted taxonomies.}

\item{domain_name}{String specifying the domain name to use for all species. The IUCN Red List files do not include domain information, so a domain name must be provided. If using a reference database from UNITE, provide a kingdom name here (e.g., \code{'Fungi'}). The default is \code{'Eukaryota'}.}

\item{path_to_taxonomy_edits}{String specifying path to taxonomy edits file in CSV format (with '.csv' extension). The file must contain the following fields: 'Old_Taxonomy', 'New_Taxonomy', 'Notes'. Old taxonomies are replaced with new taxonomies in the order the records appear in the file. The taxonomic levels in the 'Old_Taxonomy' and 'New_Taxonomy' fields should be delimited by a semi-colon. If no taxonomy edits are desired, then set this variable to \code{NA} (the default).}
}
\description{
Formats taxonomies from IUCN Red List taxonomy.csv and common_names.csv files for use with the \code{local_taxa_tool} function.
}
