% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcor.R
\name{mlcor}
\alias{mlcor}
\title{Correlations of Multivariate Logistic Regression Models}
\usage{
mlcor(fit, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), dimnames)
}
\arguments{
\item{fit}{A \code{stanfit} object returned from the \code{\link[=mlreg]{mlreg}} function. The fitted multivariate logistic regression model.}

\item{probs}{Numeric vector of probabilities. Passed to the \code{probs} argument of the \code{\link[stats:quantile]{stats::quantile}} function. The length of \code{probs} defines the length of the first dimension of the returned 3-dimensional array. By default, \code{probs = c(0.025, 0.25, 0.5, 0.75, 0.975)}. See details for the interpretation of values at each probability.}

\item{dimnames}{List (optional). If provided, then names within the returned 3-dimensional array will receive these values. Passed to the \code{dimnames} argument of the \code{\link[=array]{array}} function. If omitted, then generic names will be provided to the returned 3D array. See the \code{dimnames} argument of the \code{\link[=array]{array}} function for details.}
}
\value{
Numeric 3-dimensional array of residual correlation posterior quantiles.
}
\description{
Extract residual correlations from multivariate logistic regression models.
}
\details{
Extracts residual correlation estimates from a multivariate logistic regression model fit using the \code{\link[=mlreg]{mlreg}} function. Summarizes estimates by the quantiles of their posterior distributions, and returns summaries in a 3-dimensional array. The dimensions of the 3D array represent the posterior quantiles (dimension 1) and the response variables (both dimensions 2 and 3). Values at \code{probs = 0.025} and \code{0.975} comprise 95\% credible intervals. Values at \code{probs = 0.25} and \code{0.75} comprise 50\% credible intervals, and values at \code{probs = 0.5} represent point estimates.
}
\examples{
# Define example data file path.
path<-system.file("extdata",
                  "example_mvlogistic_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Extract residual correlations.
out<-mlcor(fit=data$fit)
}
\seealso{
\code{\link[=mlreg]{mlreg}} for fitting multivariate logistic regression models. \cr \cr
\code{\link[=mlcoef]{mlcoef}} for extracting regression coefficients from multivariate logistic regression models. \cr \cr
\code{\link[=mlformat]{mlformat}} for formatting output of multivariate logistic regression models.
}
