% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpredict.R
\name{mlpredict}
\alias{mlpredict}
\title{Predictions for Multivariate Logistic Regression Models}
\usage{
mlpredict(X, fit, names)
}
\arguments{
\item{X}{Numeric predictor matrix. Predictions are made for each record. Each field represents a predictor variable, and the predictor variables must match (in order) those used to fit the \code{\link[=mlreg]{mlreg}} model. Matrix cells contain predictor values. Element names in the returned list are taken from the row names of \code{X}.}

\item{fit}{A \code{stanfit} object returned from the \code{\link[=mlreg]{mlreg}} function. The fitted multivariate logistic regression model.}

\item{names}{Vector (optional). If provided, then field names in the matrices of the returned list will receive these values. If omitted, then the matrices in the returned list will lack field names.}
}
\value{
A list whose elements contain numeric matrices of posterior predictions. Within the list, one element is returned for each record of \code{X}. Element names are taken from the row names of \code{X}.
}
\description{
Generate predictions for multivariate logistic regression models.
}
\details{
Generates posterior predictions for multivariate logistic regression models fit with the \code{\link[=mlreg]{mlreg}} function. Returns a list where each element contains a matrix of posterior predictions for the respective record of \code{X}. Field names for the element matrices can optionally be provided with the \code{names} argument.
}
\examples{
# Define example data file path.
path<-system.file("extdata",
                  "example_mvlogistic_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Predict with fitted multivariate logistic regression.
out<-mlpredict(X=data$X,fit=data$fit,names=colnames(data$Y))
}
\seealso{
\code{\link[=mlreg]{mlreg}} for fitting multivariate logistic regression models. \cr \cr
\code{\link[=mlformat]{mlformat}} for formatting output of multivariate logistic regression models. \cr \cr
\code{\link[=mlWAIC]{mlWAIC}} for computing widely applicable information criteria for multivariate logistic regression models.
}
