% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlreg.R
\name{mlreg}
\alias{mlreg}
\title{Fitting Multivariate Logistic Regression Models}
\usage{
mlreg(
  Y,
  X,
  priors = c(B.mu = 0, B.sd = 1, lkj = 1),
  iter = 20000,
  thin = 20,
  control = list(adapt_delta = 0.99, max_treedepth = 20, stepsize = 0.01),
  ...
)
}
\arguments{
\item{Y}{Numeric response matrix. Each record represents an observation, and each field represents a response dimension. Matrix cells contain binary values (\emph{i.e.}, \code{0} or \code{1}).}

\item{X}{Numeric predictor matrix. Each record represents an observation, and each field represents a predictor variable. Matrix cells contain predictor values.}

\item{priors}{Named numeric vector. Elements represent the prior values of their respective named parameters. When predictors are centered and scaled, the defaults generally represent weakly informative priors. Regression coefficients (\code{B}) receive normal priors (with standard normal as the default). The residual correlation matrix (\code{R}) receives an LKJ prior (with default shape parameter of \code{1}).}

\item{iter}{Numeric scalar. Integer value specifying the number of iterations for each chain (including warmup). The default is \code{20000}. Passed to the \code{iter} argument of the \code{rstan::sampling} function.}

\item{thin}{Numeric scalar. Integer value specifying the thinning interval. The default is \code{20}. Passed to the \code{thin} argument of the \code{rstan::sampling} function.}

\item{control}{Named list of parameters which control the behavior of the Stan sampler. Passed to the \code{control} argument of the \code{rstan::sampling} function.}

\item{...}{Additional arguments passed to the \code{rstan::sampling} function.}
}
\value{
Returns a \code{stanfit} object of the fitted multivariate logistic regression model.
}
\description{
Fit a multivariate logistic regression model. Installation of the \code{rstan} package is required to use this function.
}
\details{
Fits a multivariate logistic regression model using the \code{rstan} interface to Stan (Carpenter \emph{et al.} 2017). The multivariate logistic regression follows that of Ovaskainen \emph{et al.} 2010, where the Bernoulli marginals are reparameterized as truncated continuous latent variables (Albert & Chib 1993). The latent variables \code{z} receive a positive constraint when \code{y = 1} and a negative constraint when \code{y = 0}, where \code{z} is a linear combination of predictors with correlated standard logistic errors. Equivalently, the latent variables follow a multivariate logistic distribution with scale parameters fixed at one (O'Brien & Dunson 2004), constructed in Stan as a Gaussian copula with logistic marginals (Song 2000). A \code{stanfit} object of the fitted model is returned, which can be used with standard \code{rstan} functions to evaluate model convergence (\emph{e.g.}, posterior trace plots, R-hat convergence diagnostics, and effective sample sizes). By default, weakly informative priors are used on the regression coefficients (\code{B}) and residual correlation matrix (\code{R}).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Define example data file path.
path<-system.file("extdata",
                  "example_mvlogistic_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Fit multivariate logistic regression.
out<-mlreg(Y=data$Y,X=data$X)
\dontshow{\}) # examplesIf}
}
\references{
Albert JH, and Chib S. 1993. Bayesian analysis of binary and polychotomous response data. \emph{Journal of the American Statistical Association}, 88(422): 669-679. \cr \cr
Carpenter B, Gelman A, Hoffman MD, Lee D, Goodrich B, Betancourt M, Brubaker M, Guo J, Li P, and Riddell A. 2017. Stan: A probabilistic programming language. \emph{Journal of Statistical Software}, 76: 1-32. DOI: 10.18637/jss.v076.i01 \cr \cr
O'Brien SM, and Dunson DB. 2004. Bayesian multivariate logistic regression. \emph{Biometrics}, 60: 739-746. DOI: 10.1111/j.0006-341X.2004.00224.x \cr \cr
Ovaskainen O, Hottola J, and Siitonen J. 2010. Modeling species co-occurrence by multivariate logistic regression generates new hypotheses on fungal interactions. \emph{Ecology}, 91(9): 2514-2521. DOI: 10.1890/10-0173.1 \cr \cr
Song P. 2000. Multivariate dispersion models generated from Gaussian copula. \emph{Scandinavian Journal of Statistics}, 27(2): 305-320. DOI: 10.1111/1467-9469.00191
}
\seealso{
\code{\link[=mlpredict]{mlpredict}} for generating predictions from multivariate logistic regression models. \cr \cr
\code{\link[=mlWAIC]{mlWAIC}} for computing widely applicable information criteria for multivariate logistic regression models.
}
