% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlWAIC.R
\name{mlWAIC}
\alias{mlWAIC}
\title{WAIC for Multivariate Logistic Regression Models}
\usage{
mlWAIC(
  Y,
  X,
  multivariate = TRUE,
  method = 2,
  priors = c(B.mu = 0, B.sd = 1, lkj = 1),
  iter = 20000,
  thin = 20,
  control = list(adapt_delta = 0.99, max_treedepth = 20, stepsize = 0.01),
  ...
)
}
\arguments{
\item{Y}{Numeric response matrix. Each record represents an observation, and each field represents a response dimension. Matrix cells contain binary values (\emph{i.e.}, \code{0} or \code{1}).}

\item{X}{Numeric predictor matrix. Each record represents an observation, and each field represents a predictor variable. Matrix cells contain predictor values.}

\item{multivariate}{Logical scalar. If \code{TRUE} (the default), then fits a multivariate logistic regression. If \code{FALSE}, then fits stacked univariate logistic regressions.}

\item{method}{Numeric scalar. Options are \code{1} or \code{2}, representing the alternative WAIC bias correction formulas (\emph{p}WAIC1 and \emph{p}WAIC2, respectively) described in Gelman \emph{et al.} (2014). As recommended by Gelman \emph{et al.} (2014), the default \code{method} (\code{2}) uses the \emph{p}WAIC2 bias correction formula.}

\item{priors}{Named numeric vector. Elements represent the prior values of their respective named parameters. When predictors are centered and scaled, the defaults generally represent weakly informative priors. Regression coefficients (\code{B}) receive normal priors (with standard normal as the default). The residual correlation matrix (\code{R}) receives an LKJ prior (with default shape parameter of \code{1}).}

\item{iter}{Numeric scalar. Integer value specifying the number of iterations for each chain (including warmup). The default is \code{20000}. Passed to the \code{iter} argument of the \code{rstan::sampling} function.}

\item{thin}{Numeric scalar. Integer value specifying the thinning interval. The default is \code{20}. Passed to the \code{thin} argument of the \code{rstan::sampling} function.}

\item{control}{Named list of parameters which control the behavior of the Stan sampler. Passed to the \code{control} argument of the \code{rstan::sampling} function.}

\item{...}{Additional arguments passed to the \code{rstan::sampling} function.}
}
\value{
Returns numeric scalar of the widely applicable information criterion.
}
\description{
Computes the widely applicable information criterion (WAIC) for multivariate logistic regression models. Serves as a wrapper for \code{\link[=mlreg]{mlreg}}, \code{\link[=mlpredict]{mlpredict}}, \code{\link[=djsdm]{djsdm}}, and \code{\link[=waic]{waic}} for convenient WAIC calculations. Installation of the \code{rstan} package is required to use this function.
}
\details{
For convenience, wraps the steps involved in WAIC calculations for Bayesian multivariate logistic regression models. Begins by fitting a Bayesian multivariate logistic regression model with the \code{\link[=mlreg]{mlreg}} function, then generates resubstitution posterior predictions using the \code{\link[=mlpredict]{mlpredict}} function. The pointwise log-likelihood is calculated with the \code{\link[=djsdm]{djsdm}} function given the response matrix and posterior predictions. WAIC is calculated from the pointwise log-likelihood using the \code{\link[=waic]{waic}} function. Because \code{\link[=djsdm]{djsdm}} does not consider residual correlations in density calculations, species interactions do not contribute to WAIC (\emph{i.e.}, response dimensions are independent).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Define example data file path.
path<-system.file("extdata",
                  "example_mvlogistic_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Compute WAIC for multivariate logistic regression.
out<-mlWAIC(Y=data$Y,X=data$X)
\dontshow{\}) # examplesIf}
}
\references{
Carpenter B, Gelman A, Hoffman MD, Lee D, Goodrich B, Betancourt M, Brubaker M, Guo J, Li P, and Riddell A. 2017. Stan: A probabilistic programming language. \emph{Journal of Statistical Software}, 76: 1-32. DOI: 10.18637/jss.v076.i01 \cr \cr
Gelman A, Hwang J, and Vehtari A. 2014. Understanding predictive information criteria for Bayesian models. \emph{Statistics and Computing}, 24(6): 997-1016. DOI: 10.1007/s11222-013-9416-2 \cr \cr
O'Brien SM, and Dunson DB. 2004. Bayesian multivariate logistic regression. \emph{Biometrics}, 60: 739-746. DOI: 10.1111/j.0006-341X.2004.00224.x \cr \cr
Ovaskainen O, Hottola J, and Siitonen J. 2010. Modeling species co-occurrence by multivariate logistic regression generates new hypotheses on fungal interactions. \emph{Ecology}, 91(9): 2514-2521. DOI: 10.1890/10-0173.1 \cr \cr
Watanabe S. 2010. Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory. \emph{Journal of Machine Learning Research}, 11(116): 3571-3594.
}
\seealso{
\code{\link[=mlreg]{mlreg}} for fitting multivariate logistic regression models. \cr \cr
\code{\link[=mlpredict]{mlpredict}} for generating predictions from multivariate logistic regression models. \cr \cr
\code{\link[=djsdm]{djsdm}} for probability mass function of a joint species distribution model. \cr \cr
\code{\link[=waic]{waic}} for generic function to compute widely applicable information criterion.
}
