% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a Vector or Matrix}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric vector or matrix. If vector, then the vector will be normalized to sum to one. If matrix, then each record will be normalized to sum to one (and a matrix returned).}
}
\value{
A numeric vector or matrix whose elements (if vector) or records (if matrix) sum to one.
}
\description{
Normalizes a vector or each record of a matrix into a simplex.
}
\details{
Returns a vector or matrix whose elements (if vector) or records (if matrix) are computed as \verb{x/}\code{\link[=sum]{sum}}\code{(x)}. This normalizes a vector or matrix record into a set of proportions which sum to one (\emph{i.e.}, a simplex). If a matrix is provided for the \code{x} argument, then normalization is performed independently for each record.
}
\examples{
# Normalize vector.
normalize(x=c(3,1,5,7))
}
\seealso{
\code{\link[=softmax]{softmax}} for the softmax function.
}
