% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Grouped Proportion Plot}
\usage{
proportion(
  x,
  r = 1,
  b = 0.025,
  v = 1000,
  w = 1,
  f = 0.5,
  c = "lightskyblue",
  s = FALSE,
  a = TRUE,
  m = 3,
  ...
)
}
\arguments{
\item{x}{A list of vectors named \code{"g"}, \code{"s"}, and \code{"p"}. The elements of vector \code{"g"} (character or numeric) specify the group. The elements of vector \code{"s"} (character or numeric) specify the sample. The elements of vector \code{"p"} (numeric) specify the proportional abundance within sample \code{"s"}.}

\item{r}{Numeric scalar. Radius of plot circle (default = \code{1}).}

\item{b}{Numeric scalar. Plot radius buffer (proportion; default = \code{0.025}).}

\item{v}{Numeric scalar. Vertex count of plot circle (default = \code{1000}).}

\item{w}{Numeric scalar. Line width of outer circle (default = \code{1}).}

\item{f}{Numeric scalar. Line width of sectors as a proportion of \code{w} (default = \code{0.5}).}

\item{c}{Character string. Fill color of sector proportions (default = \code{"lightskyblue"}).}

\item{s}{Logical value. If \code{FALSE} (the default), sort samples alphabetically. If \code{TRUE}, sort samples by decreasing proportional abundance. Samples are sorted independently for each group.}

\item{a}{Logical value. If \code{FALSE}, proportional abundance is represented by the fraction of filled sector radius to outer sector radius. If \code{TRUE} (the default), proportional abundance is represented by the fraction of filled sector area to outer sector area.}

\item{m}{Numeric scalar. Maximum number of plot columns (default = \code{3}).}

\item{...}{Additional arguments passed to \code{\link[graphics:title]{title}}.}
}
\value{
No return value.
}
\description{
Generates proportion plots for multiple groups.
}
\details{
Produces a pie-chart-like proportion plot with grouping structure. Each circle represents a group, and each sector represents a sample. When \code{a} = \code{TRUE} (the default), then the proportion of each sector filled with color represents the within-sample proportional abundance. Groups are sorted alphabetically and arranged from left to right and top to bottom. When \code{s} = \code{FALSE} (the default), then samples are sorted alphabetically and arranged in a clockwise orientation (from angle zero). When \code{s} = \code{TRUE}, then samples are sorted by decreasing proportional abundance. Samples are sorted independently for each group. This plot design is specialized for visualizing proportional abundance data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set.seed(1234)
n.groups<-6
n.samples<-6
data<-list(g=rep(x=LETTERS[1:n.groups],each=n.samples),
           s=rep(x=letters[1:n.samples],times=n.groups),
           p=stats::rbeta(n=n.groups*n.samples,
                          shape1=1,shape2=1))
proportion(x=data)
\dontshow{\}) # examplesIf}
}
\references{
A manuscript describing this plot design is in preparation.
}
\seealso{
\code{\link[=singular.proportion]{singular.proportion}} for singular proportion plots. \cr \cr
\code{\link[=detection]{detection}} for grouped detection plots.
}
