% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complement.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Get the Reverse Complement of a DNA Sequence}
\usage{
reverse_complement(sequence)
}
\arguments{
\item{sequence}{A string specifying the DNA sequence. Can contain ambiguous nucleotides.}
}
\value{
A string of the reverse complement of the DNA sequence.
}
\description{
Gets the reverse complement of a DNA sequence. Ambiguous nucleotides are supported.
}
\examples{
reverse_complement(sequence="TTCTCCASCCGCGGATHTTG")
}
