% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{richness}
\alias{richness}
\title{Species Richness}
\usage{
richness(psi)
}
\arguments{
\item{psi}{Numeric vector or matrix of occupancy probabilities. If vector, then species richness is computed for the vector of probabilities (and a scalar is returned). If matrix, then species richness is computed independently for each record (and a vector is returned).}
}
\value{
Numeric scalar or vector of species richness values.
}
\description{
Compute species richness from occupancy probabilities.
}
\details{
Calculates species richness from occupancy probabilities. Given a vector of species occupancy probabilities, computes the expected number of species as the sum of the probabilities. If given a matrix of species occupancy probabilities (where each record represents a community), computes the expected number of species as the row sums.
}
\examples{
# Compute species richness.
richness(psi=c(0.506,0.825,0.135,0.683))
}
\seealso{
\code{\link[=diversity]{diversity}} for computing Hill diversity from proportional abundances. \cr \cr
\code{\link[=dissimilarity]{dissimilarity}} for computing Bray-Curtis dissimilarity from proportional abundances.
}
