% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{sector}
\alias{sector}
\title{Draw Sector Polygon}
\usage{
sector(s, e, r, v = 1000, ...)
}
\arguments{
\item{s}{Numeric scalar of start angle (degrees).}

\item{e}{Numeric scalar of end angle (degrees).}

\item{r}{Numeric scalar of circle radius.}

\item{v}{Numeric scalar of full-circle vertex count (default = \code{1000}).}

\item{...}{Additional arguments passed to \code{\link[graphics:polygon]{polgyon}}.}
}
\value{
No return value.
}
\description{
Draws sector polygon.
}
\details{
Draws a sector polygon given a start angle, end angle, and circle radius. The sector is drawn about the origin (\emph{i.e.}, x = 0, y = 0). Intended for use with \code{\link[=template]{template}} to generate \code{\link[=detection]{detection}} and \code{\link{proportion}} plots.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
template(l=1)
sector(s=0,e=45,r=1)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=circle]{circle}} for plotting circle polygons.
}
