% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singular.detection.R
\name{singular.detection}
\alias{singular.detection}
\title{Singular Detection Plot}
\usage{
singular.detection(
  x,
  r = 1,
  b = 0.025,
  v = 1000,
  w = 1,
  f = 0.5,
  c = "lightskyblue",
  t = "",
  ...
)
}
\arguments{
\item{x}{A list of vectors named \code{"s"}, \code{"r"}, and \code{"d"}. The elements of vector \code{"s"} (character or numeric) specify the sample. The elements of vector \code{"r"} (numeric) specify the number of replicates within sample \code{"s"}. The elements of vector \code{"r"} (numeric) specify the number of replicates within sample \code{"s"} with detections.}

\item{r}{Numeric scalar. Radius of plot circle (default = \code{1}).}

\item{b}{Numeric scalar. Plot radius buffer (proportion; default = \code{0.025}).}

\item{v}{Numeric scalar. Vertex count of plot circle (default = \code{1000}).}

\item{w}{Numeric scalar. Line width of outer circle (default = \code{1}).}

\item{f}{Numeric scalar. Line width of sectors as a proportion of \code{w} (default = \code{0.5}).}

\item{c}{Character string. Fill color of sub-sector detections (default = \code{"lightskyblue"}).}

\item{t}{Character string. Plot title (default = \code{""}).}

\item{...}{Additional arguments passed to \code{\link[graphics:title]{title}}.}
}
\value{
No return value.
}
\description{
Generates a detection plot for a singular group.
}
\details{
Produces a pie-chart-like detection plot without grouping structure. Each sector represents a sample, and each sub-sector represents a replicate. Filled replicates represent detections. Samples are sorted alphabetically and arranged in a clockwise orientation (from angle zero). This plot design is specialized for visualizing binary detection data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set.seed(1234)
n.samples<-6
n.replicates<-3
data<-list(s=letters[1:n.samples],
           r=rep(x=n.replicates,times=n.samples),
           d=sample(x=0:n.replicates,size=n.samples,
                    replace=TRUE))
singular.detection(x=data)
\dontshow{\}) # examplesIf}
}
\references{
A manuscript describing this plot design is in preparation.
}
\seealso{
\code{\link[=detection]{detection}} for grouped detection plots. \cr \cr
\code{\link[=proportion]{proportion}} for grouped proportion plots.
}
