% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singular.proportion.R
\name{singular.proportion}
\alias{singular.proportion}
\title{Singular Proportion Plot}
\usage{
singular.proportion(
  x,
  r = 1,
  b = 0.025,
  v = 1000,
  w = 1,
  f = 0.5,
  c = "lightskyblue",
  t = "",
  s = FALSE,
  a = TRUE,
  ...
)
}
\arguments{
\item{x}{A list of vectors named \code{"s"} and \code{"p"}. The elements of vector \code{"s"} (character or numeric) specify the sample. The elements of vector \code{"p"} (numeric) specify the proportional abundance within sample \code{"s"}.}

\item{r}{Numeric scalar. Radius of plot circle (default = \code{1}).}

\item{b}{Numeric scalar. Plot radius buffer (proportion; default = \code{0.025}).}

\item{v}{Numeric scalar. Vertex count of plot circle (default = \code{1000}).}

\item{w}{Numeric scalar. Line width of outer circle (default = \code{1}).}

\item{f}{Numeric scalar. Line width of sectors as a proportion of \code{w} (default = \code{0.5}).}

\item{c}{Character string. Fill color of sector proportions (default = \code{"lightskyblue"}).}

\item{t}{Character string. Plot title (default = \code{""}).}

\item{s}{Logical value. If \code{FALSE} (the default), sort samples alphabetically. If \code{TRUE}, sort samples by decreasing proportional abundance.}

\item{a}{Logical value. If \code{FALSE}, proportional abundance is represented by the fraction of filled sector radius to outer sector radius. If \code{TRUE} (the default), proportional abundance is represented by the fraction of filled sector area to outer sector area.}

\item{...}{Additional arguments passed to \code{\link[graphics:title]{title}}.}
}
\value{
No return value.
}
\description{
Generates a proportion plot for a singular group.
}
\details{
Produces a pie-chart-like proportion plot without grouping structure. Each sector represents a sample. When \code{a} = \code{TRUE} (the default), then the proportion of each sector filled with color represents the within-sample proportional abundance. When \code{s} = \code{FALSE} (the default), then samples are sorted alphabetically and arranged in a clockwise orientation (from angle zero). When \code{s} = \code{TRUE}, then samples are sorted by decreasing proportional abundance. This plot design is specialized for visualizing proportional abundance data.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
set.seed(1234)
n.samples<-6
data<-list(s=letters[1:n.samples],
           p=stats::rbeta(n=n.samples,
                          shape1=1,shape2=1))
singular.proportion(x=data)
\dontshow{\}) # examplesIf}
}
\references{
A manuscript describing this plot design is in preparation.
}
\seealso{
\code{\link[=proportion]{proportion}} for grouped proportion plots. \cr \cr
\code{\link[=singular.detection]{singular.detection}} for singular detection plots.
}
