% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{The Softmax}
\usage{
softmax(x)
}
\arguments{
\item{x}{Numeric vector or matrix. If vector, then the softmax of the vector will be returned. If matrix, then the softmax will be applied independently to each record (and a matrix returned).}
}
\value{
A numeric vector or matrix whose elements (if vector) or records (if matrix) sum to one.
}
\description{
Applies the softmax to a vector or each record of a matrix.
}
\details{
Returns a vector or matrix whose elements (if vector) or records (if matrix) are computed as \code{\link[=exp]{exp}}\verb{(x)/}\code{\link[=sum]{sum}}\code{(}\code{\link[=exp]{exp}}\verb{(x))}. The softmax converts a vector or matrix record into a set of proportions which sum to one. If a matrix is provided for the \code{x} argument, then the softmax is applied independently for each record.
}
\examples{
# Perform softmax on vector.
softmax(x=c(-0.25,0.75,1.5,0))
}
\seealso{
\code{\link[=normalize]{normalize}} for vector or matrix normalization.
}
