% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_wildcards.R
\name{substitute_wildcards}
\alias{substitute_wildcards}
\title{Substitute Wildcard Characters in a DNA Sequence}
\usage{
substitute_wildcards(sequence)
}
\arguments{
\item{sequence}{A string specifying the DNA sequence containing wildcard characters.}
}
\value{
A string of the DNA sequence in which wildcard characters are replaced with their associated nucleotides surrounded by square brackets.
}
\description{
Substitutes wildcard characters in a DNA sequence with their associated nucleotides surrounded by square brackets. The output is useful for matching in regular expressions.
}
\examples{
substitute_wildcards(sequence="CAADATCCGCGGSTGGAGAA")
}
