% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_quality_scores.R
\name{summarize_quality_scores}
\alias{summarize_quality_scores}
\title{Summarize Quality Scores}
\usage{
summarize_quality_scores(
  forward_files,
  reverse_files,
  n.total = 10000,
  n.each = ceiling(n.total/length(forward_files)),
  seed = NULL,
  FUN = mean,
  ...
)
}
\arguments{
\item{forward_files}{A character vector of file paths to FASTQ files containing forward DNA sequence reads.}

\item{reverse_files}{A character vector of file paths to FASTQ files containing reverse DNA sequence reads.}

\item{n.total}{Numeric. The number of read pairs to randomly sample from the input FASTQ files. Ignored if \code{n.each} is specified. The default is \code{10000}.}

\item{n.each}{Numeric. The number of read pairs to randomly sample from each pair of input FASTQ files. The default is \code{ceiling(n.total/length(forward_files))}.}

\item{seed}{Numeric. The seed for randomly sampling read pairs. If \code{NULL} (the default), then a random seed is used.}

\item{FUN}{A function to compute summary statistics of the quality scores. The default is \code{\link[=mean]{mean}}.}

\item{...}{Additional arguments passed to \code{FUN}.}
}
\value{
Returns a data frame containing summary statistics of read length and quality score at each base pair position. The returned data frame contains the following fields:
\itemize{
\item Direction: The read direction (\emph{i.e.}, "Forward" or "Reverse").
\item Position: The base pair position.
\item Length: The summary statistic(s) of read lengths. If \code{FUN} returns multiple summary statistics, then a matrix of the summary statistics will be stored in this field, which can be accessed with \code{\link[=Extract]{$Length}}.
\item Score: The summary statistic(s) of Phred quality scores. If \code{FUN} returns multiple summary statistics, then a matrix of the summary statistics will be stored in this field, which can be accessed with \code{\link[=Extract]{$Score}}.
\item Probability: The summary statistic(s) of the cumulative probability that all bases were called correctly. If \code{FUN} returns multiple summary statistics, then a matrix of the summary statistics will be stored in this field, which can be accessed with \code{\link[=Extract]{$Probability}}.
}
}
\description{
For each base pair position, summarizes read length, Phred quality score, and the cumulative probability that all bases were called correctly.
}
\details{
For each combination of base pair position and read direction, calculates summary statistics of read length, Phred quality score, and the cumulative probability that all bases were called correctly. The cumulative probability is calculated from the first base pair up to the current position. Quality scores are assumed to be encoded in Sanger format. Read pairs are selected by randomly sampling up to \code{n.each} read pairs from each pair of input FASTQ files. By default, \code{n.each} is derived from \code{n.total}, and \code{n.total} will be ignored if \code{n.each} is provided. By default, \code{\link[=mean]{mean}} is used to compute the summary statistics, but the user may provide another summary function instead (\emph{e.g.}, \code{\link[stats:median]{median}}). Functions which return multiple summary statistics are also supported (\emph{e.g.}, \code{\link[=summary]{summary}} and \code{\link[stats:quantile]{quantile}}). Arguments in \code{...} are passed to the summary function.
}
\examples{
# Get example forward FASTQ files.
forward_files<-system.file("extdata",
                           paste0("S0",1:3,"F.fastq"),
                           package="LocaTT",
                           mustWork=TRUE)

# Get example reverse FASTQ files.
reverse_files<-system.file("extdata",
                           paste0("S0",1:3,"R.fastq"),
                           package="LocaTT",
                           mustWork=TRUE)

# Summarize quality scores.
summarize_quality_scores(forward_files,reverse_files)
}
\seealso{
\code{\link[=decode_quality_scores]{decode_quality_scores}} for decoding quality scores.
}
