% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Initiate Template Plot}
\usage{
template(l, b = 0.025)
}
\arguments{
\item{l}{Numeric scalar of axis limits (applies to both axes).}

\item{b}{Numeric scalar to extend axis limits (see Details; default = \code{0.025}).}
}
\value{
No return value.
}
\description{
Initiates a blank template plot.
}
\details{
Initiates a blank template plot with limits \code{l} and buffer \code{b} about the origin (\emph{i.e.}, x = 0, y = 0). \code{l} is used for axis limits in both the negative and positive directions. \code{b} extends the limits beyond \code{l} by a fixed proportion (\emph{i.e.}, \code{l} * (1 + \code{b})). Intended for use with \code{\link[=circle]{circle}} and \code{\link[=sector]{sector}}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
template(l=1)
circle(r=1)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=circle]{circle}} for plotting circle polygons. \cr \cr
\code{\link[=sector]{sector}} for plotting sector polygons.
}
