% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_sequences.R
\name{trim_sequences}
\alias{trim_sequences}
\title{Trim Target Nucleotide Sequence from DNA Sequences}
\usage{
trim_sequences(
  sequences,
  target,
  anchor = "start",
  required = TRUE,
  quality_scores
)
}
\arguments{
\item{sequences}{A character vector of DNA sequences to trim.}

\item{target}{A string specifying the target nucleotide sequence.}

\item{anchor}{A string specifying whether the target nucleotide sequence should be trimmed from the start or end of the DNA sequences. Allowable values are "start" (the default) and "end".}

\item{required}{A logical value specifying whether trimming is required. If \code{TRUE} (the default), then sequences which could not be trimmed are returned as \code{NA}s. If \code{FALSE}, then untrimmed sequences are returned along with DNA sequences for which trimming was successful.}

\item{quality_scores}{An optional character vector of DNA sequence quality scores. If supplied, these will be trimmed to their corresponding trimmed DNA sequences.}
}
\value{
If quality scores are not provided, then a character vector of trimmed DNA sequences is returned. If quality scores are provided, then a list containing two elements is returned. The first element is a character vector of trimmed DNA sequences, and the second element is a character vector of quality scores which have been trimmed to their corresponding trimmed DNA sequences.
}
\description{
Trims a target nucleotide sequence from the front or back of DNA sequences. Ambiguous nucleotides in the target nucleotide sequence are supported.
}
\details{
For each DNA sequence, the target nucleotide sequence is searched for at either the front or back of the DNA sequence, depending on the value of the anchor argument. If the target nucleotide sequence is found, then it is removed from the DNA sequence. If the required argument is set to \code{TRUE}, then DNA sequences in which the target nucleotide sequence was not found will be returned as \code{NA}s. If the required argument is set to \code{FALSE}, then untrimmed DNA sequences will be returned along with DNA sequences for which trimming was successful. Ambiguous nucleotides in the target nucleotide sequence are supported through the internal use of the \code{substitute_wildcards} function on the target nucleotide sequence, and a regular expression with a leading or ending anchor is used to search for the target nucleotide sequence in the DNA sequences. Trimming will fail for DNA sequences which contain ambiguous nucleotides (\emph{e.g.}, Ns) in their target nucleotide sequence region, resulting in \code{NA}s for those sequences if the required argument is set to \code{TRUE}.
}
\examples{
trim_sequences(sequences=c("ATATAGCGCG","TGCCGATATA","ATCTATCACCGC"),
               target="ATMTA",
               anchor="start",
               required=TRUE,
               quality_scores=c("989!.C;F@\"","A((#-#;,2F","HD8I/+67=1>?"))
}
