% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_and_merge_pairs.R
\name{truncate_and_merge_pairs}
\alias{truncate_and_merge_pairs}
\title{Truncate and Merge Forward and Reverse DNA Sequence Reads}
\usage{
truncate_and_merge_pairs(
  forward_files,
  reverse_files,
  output_files,
  truncation_length = NA,
  threshold.quality_score = 3,
  threshold.probability = 0.5,
  minimum_overlap = 10,
  cores = 1,
  progress = FALSE
)
}
\arguments{
\item{forward_files}{A character vector of file paths to FASTQ files containing forward DNA sequence reads.}

\item{reverse_files}{A character vector of file paths to FASTQ files containing reverse DNA sequence reads.}

\item{output_files}{A character vector of file paths to output FASTA files.}

\item{truncation_length}{Numeric. The length to truncate DNA sequences to (passed to the \code{length} argument of \code{\link[=truncate_sequences.length]{truncate_sequences.length}}). If \code{NA} (the default), then DNA sequences are not truncated by length. If a single value is supplied, then both forward and reverse reads are truncated to the same length. If two values are supplied in a numeric vector, then the first value is used to truncate the forward reads, and the second value is used to truncate the reverse reads. \code{NA} can also be supplied as either the first or second element of the numeric vector to prevent length truncation of the respective read direction while allowing the other read direction to be length truncated.}

\item{threshold.quality_score}{Numeric. The Phred quality score threshold used for truncation (passed to the \code{threshold} argument of \code{\link[=truncate_sequences.quality_score]{truncate_sequences.quality_score}}). The default is \code{3} (\emph{i.e.}, each base in a truncated sequence has a greater than 50\% probability of having been called correctly). If \code{NA}, then DNA sequences are not truncated by quality score threshold. If a single value is supplied, then both forward and reverse reads are truncated by the same quality score threshold. If two values are supplied in a numeric vector, then the first value is used to truncate the forward reads, and the second value is used to truncate the reverse reads. \code{NA} can also be supplied as either the first or second element of the numeric vector to prevent quality-score-threshold truncation of the respective read direction while allowing the other read direction to be quality-score-threshold truncated.}

\item{threshold.probability}{Numeric. The probability threshold used for truncation (passed to the \code{threshold} argument of \code{\link[=truncate_sequences.probability]{truncate_sequences.probability}}). The default is \code{0.5} (\emph{i.e.}, each truncated sequence has a greater than 50\% probability that all bases were called correctly). If \code{NA}, then DNA sequences are not truncated by probability threshold. If a single value is supplied, then both forward and reverse reads are truncated by the same probability threshold. If two values are supplied in a numeric vector, then the first value is used to truncate the forward reads, and the second value is used to truncate the reverse reads. \code{NA} can also be supplied as either the first or second element of the numeric vector to prevent probability-threshold truncation of the respective read direction while allowing the other read direction to be probability-threshold truncated.}

\item{minimum_overlap}{Numeric. The minimum length of an overlap that must be found between the end of the forward read and the start of the reverse complement of the reverse read in order for a read pair to be merged (passed to \code{\link[=merge_pairs]{merge_pairs}}). The default is \code{10}.}

\item{cores}{Numeric. If \code{1} (the default), then FASTQ file pairs are processed sequentially on a single core. If greater than \code{1}, then FASTQ file pairs are processed in parallel across the specified number of cores. Parallel processing is not supported on Windows.}

\item{progress}{Logical. If \code{TRUE}, then a progress indicator is printed to the console. Ignored if \code{cores > 1}. If \code{FALSE} (the default), then no progress indicator is displayed.}
}
\value{
If \code{cores = 1}, then no return value. Writes a FASTA file for each pair of input FASTQ files with DNA sequence counts stored in the header lines. If \code{cores > 1}, then also invisibly returns a list where each element contains warning or error messages associated with processing each pair of input FASTQ files. A \code{NULL} value in the returned list means that no warnings or errors were generated from processing the respective pair of FASTQ files.
}
\description{
Removes DNA read pairs containing ambiguous nucleotides, truncates reads by length and quality score, and merges forward and reverse reads.
}
\details{
For each pair of input FASTQ files, removes DNA read pairs containing ambiguous nucleotides, truncates reads by length, quality score threshold, and probability threshold (in that order), and then merges forward and reverse reads. Merged reads are summarized by frequency of occurrence and written to a FASTA file. See \code{\link[=contains_wildcards]{contains_wildcards}}, \code{\link[=truncate_sequences.length]{truncate_sequences.length}}, \code{\link[=truncate_sequences.quality_score]{truncate_sequences.quality_score}}, \code{\link[=truncate_sequences.probability]{truncate_sequences.probability}}, and \code{\link[=merge_pairs]{merge_pairs}} for methods. Quality scores are assumed to be encoded in Sanger format. Forward and reverse reads can be truncated by different thresholds (see \code{truncation_length}, \code{threshold.quality_score}, and \code{threshold.probability} arguments). \cr \cr
Multicore parallel processing is supported on Mac and Linux operating systems (not available on Windows). When \code{cores > 1} (parallel processing enabled), warnings and errors are printed to the console in addition to being invisibly returned as a list (see the return value section), and errors produced while processing a pair of FASTQ files will not interrupt the processing of other FASTQ file pairs. When \code{cores = 1}, FASTQ file pairs are processed sequentially on a single core, and errors will prevent the processing of subsequent FASTQ file pairs (but warnings will not).
}
\examples{
# Get example forward FASTQ files.
forward_files<-system.file("extdata",
                           paste0("S0",1:3,"F.fastq"),
                           package="LocaTT",
                           mustWork=TRUE)

# Get example reverse FASTQ files.
reverse_files<-system.file("extdata",
                           paste0("S0",1:3,"R.fastq"),
                           package="LocaTT",
                           mustWork=TRUE)

# Create paths for temporary output files.
output_files<-tempfile(pattern=paste0("O",1:3),fileext=".fasta")

# Truncate and merge pairs.
truncate_and_merge_pairs(forward_files=forward_files,
                         reverse_files=reverse_files,
                         output_files=output_files)
}
\references{
A manuscript describing these methods is in preparation.
}
\seealso{
\code{\link[=contains_wildcards]{contains_wildcards}} for detecting ambiguous nucleotides in DNA sequences. \cr \cr
\code{\link[=truncate_sequences.length]{truncate_sequences.length}} for truncating DNA sequences to a specified length. \cr \cr
\code{\link[=truncate_sequences.quality_score]{truncate_sequences.quality_score}} for truncating DNA sequences by Phred quality score. \cr \cr
\code{\link[=truncate_sequences.probability]{truncate_sequences.probability}} for truncating DNA sequences by cumulative probability that all bases were called correctly. \cr \cr
\code{\link[=merge_pairs]{merge_pairs}} for merging forward and reverse DNA sequence reads. \cr \cr
\code{\link[=filter_sequences]{filter_sequences}} for filtering merged read pairs by PCR replicate.
}
