% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{Widely Applicable Information Criterion}
\usage{
waic(loglik, method = 2)
}
\arguments{
\item{loglik}{Numeric matrix of the pointwise log-likelihood. Each record represents a Markov chain Monte Carlo (MCMC) sample, and each field represents an observation.}

\item{method}{Numeric scalar. Options are \code{1} or \code{2}, representing the alternative WAIC bias correction formulas (\emph{p}WAIC1 and \emph{p}WAIC2, respectively) described in Gelman \emph{et al.} (2014). As recommended by Gelman \emph{et al.} (2014), the default \code{method} (\code{2}) uses the \emph{p}WAIC2 bias correction formula.}
}
\value{
Returns numeric scalar of the widely applicable information criterion.
}
\description{
Generic function calculating widely applicable information criterion (WAIC) from the pointwise log-likelihood.
}
\details{
Given the pointwise log-likelihood, calculates WAIC (Watanabe 2010) using the formulas described in Gelman \emph{et al.} (2014). The expected log pointwise predictive density (elppd) is estimated as the log pointwise predictive density (lppd) adjusted by a bias correction (either \emph{p}WAIC1 or \emph{p}WAIC2). To reflect the deviance scale, WAIC is defined as the elppd times negative two. As recommended by Gelman \emph{et al.} (2014), \emph{p}WAIC2 is used as the default bias correction (\code{method} = \code{2}). See Gelman \emph{et al.} (2014) for details.
}
\examples{
# Define example data file path.
path<-system.file("extdata",
                  "example_regression_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Compute WAIC from pointwise log-likelihood.
out<-waic(loglik=data$loglik)
}
\references{
Gelman A, Hwang J, and Vehtari A. 2014. Understanding predictive information criteria for Bayesian models. \emph{Statistics and Computing}, 24(6): 997-1016. DOI: 10.1007/s11222-013-9416-2 \cr \cr
Watanabe S. 2010. Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory. \emph{Journal of Machine Learning Research}, 11(116): 3571-3594.
}
\seealso{
\code{\link[=ddirmult]{ddirmult}} for probability mass function of the Dirichlet-multinomial distribution. \cr \cr
\code{\link[=dmWAIC]{dmWAIC}} for computing widely applicable information criteria for Dirichlet-multinomial regression models.
}
