% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.fasta.R
\name{write.fasta}
\alias{write.fasta}
\title{Write FASTA Files}
\usage{
write.fasta(names, sequences, file)
}
\arguments{
\item{names}{A character vector of sequence names.}

\item{sequences}{A character vector of sequences.}

\item{file}{A string specifying the path to a FASTA file (with '.fasta' extension) to write.}
}
\description{
Writes FASTA files.
}
