% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_taxa_tool.R
\name{local_taxa_tool}
\alias{local_taxa_tool}
\title{Perform Geographically-Conscious Taxonomic Assignment}
\usage{
local_taxa_tool(
  path_to_sequences_to_classify,
  path_to_BLAST_database,
  path_to_output_file,
  path_to_list_of_local_taxa = NA,
  blast_e_value = 1e-05,
  blast_max_target_seqs = 2000,
  blast_task = "megablast",
  full_names = FALSE,
  separator = ", ",
  blastn_command = "blastn"
)
}
\arguments{
\item{path_to_sequences_to_classify}{String specifying path to FASTA file containing sequences to classify (with '.fasta' extension).}

\item{path_to_BLAST_database}{String specifying path to BLAST reference database in FASTA format (with '.fasta' extension).}

\item{path_to_output_file}{String specifying path to output file of classified sequences in CSV format (with '.csv' extension).}

\item{path_to_list_of_local_taxa}{String specifying path to list of local species in CSV format (with '.csv' extension). The file should contain the following fields: 'Common_Name', 'Domain', 'Phylum', 'Class', 'Order', 'Family', 'Genus', 'Species'. There should be no 'NA's or blanks in the taxonomy fields. The species field should contain the binomial name without subspecies or other information below the species level. All records in the species field should be unique. If local taxa suggestions are not desired, set this variable to \code{NA} (the default).}

\item{blast_e_value}{Numeric. Maximum E-value of returned BLAST hits (lower E-values are associated with more 'significant' matches). The default is \code{1e-05}.}

\item{blast_max_target_seqs}{Numeric. Maximum number of BLAST target sequences returned per query sequence. Enough target sequences should be returned to ensure that all minimum E-value matches are returned for each query sequence. A warning will be produced if this value is not sufficient. The default is \code{2000}.}

\item{blast_task}{String specifying BLAST task specification. Use \code{'megablast'} (the default) to find very similar sequences (e.g., intraspecies or closely related species). Use \code{'blastn-short'} for sequences shorter than 50 bases. See the blastn program help documentation for additional options and details.}

\item{full_names}{Logical. If \code{TRUE}, then full taxonomies are returned in the output CSV file. If \code{FALSE} (the default), then only the lowest taxonomic levels (e.g., species binomials instead of the full species taxonomies) are returned in the output CSV file.}

\item{separator}{String specifying the separator to use between taxa names in the output CSV file. The default is \code{', '}.}

\item{blastn_command}{String specifying path to the blastn program. The default (\code{'blastn'}) should work for standard BLAST installations. The user can provide a path to the blastn program for non-standard BLAST installations.}
}
\description{
Performs taxonomic assignment of DNA metabarcoding sequences while considering geographic location.
}
\details{
Sequences are BLASTed against a global reference database, and the tool suggests locally occurring species which are most closely related (by taxonomy) to any of the best-matching BLAST hits (by bit score). If a local taxa list is not provided, then local taxa suggestions will be disabled, but all best-matching BLAST hits will still be returned. Alternatively, a reference database containing just the sequences of local species can be used, and local taxa suggestions can be disabled to return all best BLAST matches of local species. The reference database should be formatted with the \code{format_reference_database} function, and the local taxa lists can be prepared using the \code{get_taxonomies.species_binomials} and \code{get_taxonomies.IUCN} functions. Output field definitions are:
\itemize{
\item Sequence_name: The query sequence name.
\item Sequence: The query sequence.
\item Best_match_references: Species binomials of all best-matching BLAST hits (by bit score) from the reference database.
\item Best_match_E_value: The E-value associated with the best-matching BLAST hits.
\item Best_match_bit_score: The bit score associated with the best-matching BLAST hits.
\item Best_match_query_cover.mean: The mean query cover of all best-matching BLAST hits.
\item Best_match_query_cover.SD: The standard deviation of query cover of all best-matching BLAST hits.
\item Best_match_PID.mean: The mean percent identity of all best-matching BLAST hits.
\item Best_match_PID.SD: The standard deviation of percent identity of all best-matching BLAST hits.
\item Local_taxa (Field only present if a path to a local taxa list is provided): The finest taxonomic unit(s) which include both any species of the best-matching BLAST hits and any local species. If the species of any of the best-matching BLAST hits are local, then the finest taxonomic unit(s) are at the species level.
\item Local_species (Field only present if a path to a local taxa list is provided): Species binomials of all local species which belong to the taxonomic unit(s) in the Local_taxa field.
}
}
