% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlformat.R
\name{mlformat}
\alias{mlformat}
\title{Formatting Multivariate Logistic Regression Models}
\usage{
mlformat(fit, mode = "B", ci = 0.95, digits = 3, names.x, names.y)
}
\arguments{
\item{fit}{A \code{stanfit} object returned from the \code{\link[=mlreg]{mlreg}} function. The fitted multivariate logistic regression model.}

\item{mode}{Character scalar. Specifies which parameter set to summarize. When \code{mode = "B"} (the default), generates summaries of regression coefficient estimates. When \code{mode = "R"}, generates summaries of residual correlation estimates (if \code{\link[=mlreg]{mlreg}} was fit with \code{multivariate = TRUE}).}

\item{ci}{Numeric scalar. Defines the credible interval for parameter summaries. When \code{ci = 0.95} (the default), generates 95\% credible intervals for parameter estimates. \code{ci} must be between \code{0} and \code{1}.}

\item{digits}{Numeric scalar. Positive integer value specifying the number of decimal places to which results will be rounded. The default is \code{3}.}

\item{names.x}{Character vector (optional). If provided, then supplies the names of predictor variables in the returned matrix. Names should match those of the predictor variables used to fit the \code{\link[=mlreg]{mlreg}} model.}

\item{names.y}{Character vector (optional). If provided, then supplies the names of response variables in the returned matrix. Names should match those of the response variables used to fit the \code{\link[=mlreg]{mlreg}} model.}
}
\value{
Numeric matrix of posterior summaries.
}
\description{
Format output of multivariate logistic regression models.
}
\details{
Formats output of a multivariate logistic regression model fit using the \code{\link[=mlreg]{mlreg}} function. When \code{mode = "B"} (the default), returns regression coefficient estimates. When \code{mode = "R"}, returns residual correlation estimates (if \code{\link[=mlreg]{mlreg}} was fit with \code{multivariate = TRUE}). Summarizes parameters by the quantiles of their posterior distributions, with a point estimate at the 50th percentile (\emph{i.e.}, the posterior median). Lower and upper limits are defined by the credible interval argument. At the default \code{ci = 0.95}, returns 95\% credible intervals. When a credible interval does not overlap zero, the point estimate is appended with an asterisk.
}
\examples{
# Define example data file path.
path<-system.file("extdata",
                  "example_mvlogistic_data.rds",
                  package="LocaTT",
                  mustWork=TRUE)

# Read in example regression data.
data<-readRDS(file=path)

# Retrieve fitted regression model.
fit<-data$fit

# Retrieve predictor matrix.
X<-data$X

# Retrieve response matrix.
Y<-data$Y

# Extract regression coefficients.
B<-mlformat(fit=fit,mode="B",
            names.x=colnames(X),
            names.y=colnames(Y))

# Display regression coefficients.
print(B,quote=FALSE,right=TRUE)

# Extract residual correlations.
R<-mlformat(fit=fit,mode="R",
            names.y=colnames(Y))

# Display residual correlations.
print(R,quote=FALSE,right=TRUE)
}
\seealso{
\code{\link[=mlreg]{mlreg}} for fitting multivariate logistic regression models. \cr \cr
\code{\link[=mlcoef]{mlcoef}} for extracting regression coefficients from multivariate logistic regression models. \cr \cr
\code{\link[=mlcor]{mlcor}} for extracting residual correlations from multivariate logistic regression models.
}
