% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.R
\name{read.fasta}
\alias{read.fasta}
\title{Read FASTA Files}
\usage{
read.fasta(file)
}
\arguments{
\item{file}{A string specifying the path to a FASTA file (with '.fasta' extension) to read.}
}
\value{
A data frame with fields for sequence names and sequences.
}
\description{
Reads FASTA files.
}
